Add delete dog endpoint and database actions
The API endpoint DELETE dog/[:id] has been created and calls
the database action deleteDog. This actions checks if the ID
is valid and deletes the dog if it is, returns null if the
ID does not exist, or throws an error if there is an error
with the database.