Alter some API responses and check birth order
Some of the API responses have been changed to return data and to
return an unsuccessful state when the user tries to send a request
other than a POST request. Also, now there is a check if the birth order
is greater than the litter size. Cnnecting to the database has
been put inside a try-catch block as well.