Delete Log Endpoint (#30)
The endpoint to delete logs has been added and the database actions to delete the log. Before the log is deleted, if it is present in the dog's recentLogs array, the next most recent log is found and added to the array, then the deleted log is removed from that array. Then the intended log is deleted from the database.