Dog Schema Updates (#37)
The Dog schema has been updated with more fields, including changing the partner and maternalDemeanor fields. Because of the schema change, the Zod schema for dogs has been updated as well and moved to the consts file, and the dogs/search endpoint has been updated to allow for the new partner schema. The seed file has been updated as well to reflect the dog schema changes.

Also, the loops for checking if an ID is present in the database have been changed to use the $in operator instead, and this code has been moved to a common method validateDogs.